Small warning fix

+++ src/adlmidi_midiplay.cpp
@@ -1092,7 +1092,7 @@ bool MIDIplay::realTime_NoteOn(uint8_t channel, uint8_t note, uint8_t velocity)
                         expected_mode = PercussionMap[midiins & 0xFF];
                 }
 
-                if(opl.four_op_category[a] != (uint32_t)expected_mode)
+                if(opl.four_op_category[a] != expected_mode)
                     continue;
             }
             else
